Desktop/glamai-fyp/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
anthropic>=0.91.0
bcrypt==5.0.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.2
h11==0.16.0
idna==3.11
numpy>=1.26.0
passlib==1.7.4
Pillow>=10.0.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.9
SQLAlchemy==2.0.48
starlette==0.46.2
tensorflow-macos>=2.16.0
tensorflow-metal>=1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
vaderSentiment==3.3.2
